[v4_1_esv] Update the memory leakage debug code to work with v6.

+++ omapip/alloc.c
@@ -4,6 +4,7 @@
    protocol... */
 
 /*
+ * Copyright (c) 2012 by Internet Systems Consortium, Inc. ("ISC")
  * Copyright (c) 2009-2010 by Internet Systems Consortium, Inc. ("ISC")
  * Copyright (c) 2004-2007 by Internet Systems Consortium, Inc. ("ISC")
  * Copyright (c) 1999-2003 by Internet Software Consortium
@@ -311,7 +312,7 @@ void dmalloc_dump_outstanding ()
 				inhistory = 1;
 				if (!noted) {
 				    log_info ("  %s(%d): %ld", dp -> file,
-					      dp -> line, dp -> size);
+					      dp -> line, (long) dp -> size);
 				    noted = 1;
 				}
 				print_rc_hist_entry (i);
@@ -324,7 +325,8 @@ void dmalloc_dump_outstanding ()
 			if (!inhistory)
 #endif
 				log_info ("  %s(%d): %ld",
-					  dp -> file, dp -> line, dp -> size);
+					  dp -> file, dp -> line, 
+					  (long) dp -> size);
 		}
 #endif
 	}
